Remove enkf_state.py
