+++ src/ert/callbacks.py
@@ -4,10 +4,8 @@ import logging
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, List, Tuple
 
+from ert._c_wrappers.enkf import SummaryConfig
 from ert._c_wrappers.enkf.config.parameter_config import ParameterConfig
-from ert._c_wrappers.enkf.enkf_state import (
-    _write_responses_to_storage,
-)
 from ert._c_wrappers.enkf.enums import RealizationStateEnum
 
 from .load_status import LoadResult, LoadStatus
@@ -34,6 +32,26 @@ def _read_parameters(
             error_msg += str(err)
             result = LoadResult(LoadStatus.LOAD_FAILURE, error_msg)
     return result
+
+
+def _write_responses_to_storage(
+    ens_config: EnsembleConfig, run_arg: RunArg
+) -> LoadResult:
+    errors = []
+    for config in ens_config.response_configs.values():
+        if isinstance(config, SummaryConfig):
+            # Nothing to load, should not be handled here, should never be
+            # added in the first place
+            if not config.keys:
+                continue
+        try:
+            ds = config.read_from_file(run_arg.runpath, run_arg.iens)
+            run_arg.ensemble_storage.save_response(config.name, ds, run_arg.iens)
+        except ValueError as err:
+            errors.append(str(err))
+    if errors:
+        return LoadResult(LoadStatus.LOAD_FAILURE, "\n".join(errors))
+    return LoadResult(LoadStatus.LOAD_SUCCESSFUL, "")
 
 
 def forward_model_ok(
